Update DeviceInfo.js
desired format where the first two characters are the first subnet of the IP address and the rest is the MAC address excluding the first two characters, you can modify the uniqueId computed observable in your Knockout.js component. Here is the updated code:

This code modifies the uniqueId computed observable to concatenate the first two characters of the IP address's first part with the MAC address excluding its first two characters. This should give you the result in the desired format:

Ensure that the ip_address and mac_address observables in your device observable are properly populated for the computed observables to function correctly.